Renaming PRODUCTION Environment to LIVE (#148)
**Description**

Catch.js exposes both an environment and a live setting. For the mobile SDKs, we only allow developers to operate in the production environment, but we still allow them to choose between sandbox and live.

Internally, we call sandbox and live "stages". I think it will make more sense to developers using our SDKs if we continue to call them Environments in mobile.

**Open Questions**

**Testing**

**Relevant Links (e.g. Ticket)**

- The types that Catch.js uses: https://github.com/getcatch/catchjs/blob/main/types/index.d.ts#L65-L78

**TODOs**