Close webview activity on errors (#163)
**Description**

Errors that the `WebViewClient` encounters indicate a failure to load the webpage. An example of this is when a user's device does not have any internet connection. Instead of failing to show the user anything and requiring them to hit the back button when it would not be obvious to do so, we will now automatically close the webview for them.

**Open Questions**

**Testing**

**Relevant Links (e.g. Ticket)**

**TODOs**